Added warning re outdated Clojars version
Hi - Just had some problems due to referencing the outdated `cnuernber` aritfacts on Clojars.  This name is still widely referenced in blog entries & videos and crashes when trying to run with Python 3.8, for example.  I thought a small warning on the README may help others to avoid this trap.

Thanks for a great library!

Alan